Fix Google popup bug
Adds `user.loading` field to the UserContext, which should be checked
for alongside any other "Loading" states, and *before* checking for
`user.signedIn` (which usually calls <AuthenticationRequired />,
triggering the Google login).

Defaulting to the MUI <CircularProgress /> component to indicate loading
status.

Removed <UnauthenticatedLanding /> as it was simply an alias for
<GoogleLoginButton />.